fixing warnings refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@21130 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netimport/NIXMLTrafficLightsHandler.cpp
@@ -320,7 +320,7 @@ int
 NIXMLTrafficLightsHandler::retrieveLaneIndex(
     const SUMOSAXAttributes& attrs, SumoXMLAttr attr, NBEdge* edge, bool& ok) {
     int laneIndex = attrs.get<int>(attr, 0, ok);
-    if (edge->getNumLanes() <= (size_t) laneIndex) {
+    if (edge->getNumLanes() <= laneIndex) {
         WRITE_ERROR("Invalid lane index '" + toString(laneIndex) + "' for edge '" + edge->getID() + "'.");
         ok = false;
     }
